refactor(buildSrc): migrate precompiled plugins to binary plugin
The plugin sets conventions similar to the existing precompiled script
plugins, but now allows for additional options to be set which a
precompiled script plugin would not be able to do.

Additionally, this aims to fix #7 where snapshots were accidentally
published as releases, as the check which determined if the version was
a snapshot is computed _before_ the version is actually set, resulting
in a check of `!"unspecified".endsWith("SNAPSHOT")` being `true` for
`isRelease`. This commit fixes the issue by allowing for the module's
version to be set via the plugin's extension DSL, and _only after_ the
DSL is configured do we check the version.

Fixes #7, in case GitHub doesn't detect the earlier auto-close keyword